Improved the code and the CLI command to link a chain account
