+++ proto/desmos/profiles/v1beta1/client/cli.proto
@@ -1,0 +1,34 @@
+syntax = "proto3";
+package desmos.profiles.v1beta1.client;
+
+import "gogoproto/gogo.proto";
+import "google/protobuf/any.proto";
+import "google/protobuf/timestamp.proto";
+import "cosmos_proto/cosmos.proto";
+import "desmos/profiles/v1beta1/models_chain_links.proto";
+
+option go_package = "github.com/desmos-labs/desmos/x/profiles/client/utils";
+
+// ChainLinkJSON contains the data required to create a ChainLink using the CLI
+// command
+message ChainLinkJSON {
+  option (gogoproto.equal) = true;
+  option (gogoproto.goproto_stringer) = true;
+
+  // Address contains the data of the external chain address to be connected
+  // with the Desmos profile
+  google.protobuf.Any address = 1 [
+    (gogoproto.moretags) = "yaml:\"address\"",
+    (cosmos_proto.accepts_interface) = "AddressData"
+  ];
+
+  // Proof contains the ownership proof of the external chain address
+  desmos.profiles.v1beta1.Proof proof = 2
+      [ (gogoproto.nullable) = false, (gogoproto.moretags) = "yaml:\"proof\"" ];
+
+  // ChainConfig contains the configuration of the external chain
+  desmos.profiles.v1beta1.ChainConfig chain_config = 3 [
+    (gogoproto.nullable) = false,
+    (gogoproto.moretags) = "yaml:\"chain_config\""
+  ];
+}
