+++ proto/desmos/profiles/v1beta1/models_dtag_requests.proto
@@ -5,7 +5,7 @@ import "gogoproto/gogo.proto";
 import "google/protobuf/any.proto";
 import "google/protobuf/timestamp.proto";
 import "cosmos_proto/cosmos.proto";
-import "desmos/profiles/v1beta1/models_ibc.proto";
+import "desmos/profiles/v1beta1/models_chain_links.proto";
 
 option go_package = "github.com/desmos-labs/desmos/x/profiles/types";
