+++ proto/desmos/profiles/v1beta1/models_packets.proto
@@ -4,7 +4,7 @@ package desmos.profiles.v1beta1;
 import "gogoproto/gogo.proto";
 import "google/protobuf/any.proto";
 import "cosmos_proto/cosmos.proto";
-import "desmos/profiles/v1beta1/models_ibc.proto";
+import "desmos/profiles/v1beta1/models_chain_links.proto";
 
 option go_package = "github.com/desmos-labs/desmos/x/profiles/types";
