+++ proto/desmos/profiles/v1beta1/msgs_chain_links.proto
@@ -4,35 +4,34 @@ package desmos.profiles.v1beta1;
 import "gogoproto/gogo.proto";
 import "google/protobuf/any.proto";
 import "cosmos_proto/cosmos.proto";
-import "desmos/profiles/v1beta1/models_ibc.proto";
+import "desmos/profiles/v1beta1/models_chain_links.proto";
 
 option go_package = "github.com/desmos-labs/desmos/x/profiles/types";
 
 // MsgLinkChainAccount represents a message to link an account to a profile.
 message MsgLinkChainAccount {
-  // SourceAddress contains the details of the external chain address to be
+  // ChainAddress contains the details of the external chain address to be
   // linked
-  google.protobuf.Any source_address = 1 [
+  google.protobuf.Any chain_address = 1 [
     (gogoproto.moretags) = "yaml:\"source_address\"",
     (cosmos_proto.accepts_interface) = "AddressData"
   ];
 
-  // SourceProof contains the proof of ownership of the external chain address
-  desmos.profiles.v1beta1.Proof source_proof = 2 [
+  // Proof contains the proof of ownership of the external chain address
+  desmos.profiles.v1beta1.Proof proof = 2 [
     (gogoproto.nullable) = false,
     (gogoproto.moretags) = "yaml:\"source_proof\""
   ];
 
-  // SourceChainConfig contains the configuration of the external chain
-  desmos.profiles.v1beta1.ChainConfig source_chain_config = 3 [
+  // ChainConfig contains the configuration of the external chain
+  desmos.profiles.v1beta1.ChainConfig chain_config = 3 [
     (gogoproto.nullable) = false,
     (gogoproto.moretags) = "yaml:\"source_chain_config\""
   ];
 
-  // Destination address represents the Desmos address associated with the
+  // Signer represents the Desmos address associated with the
   // profile to which link the external account
-  string destination_address = 4
-      [ (gogoproto.moretags) = "yaml:\"destination_address\"" ];
+  string signer = 4 [ (gogoproto.moretags) = "yaml:\"signer\"" ];
 }
 
 // LinkChainAccountResponse defines the Msg/LinkAccount response type.
