+++ x/profiles/client/cli/cli_relationships.go
@@ -1,0 +1,183 @@
+package cli
+
+import (
+	"context"
+	"fmt"
+
+	"github.com/cosmos/cosmos-sdk/client"
+	"github.com/cosmos/cosmos-sdk/client/flags"
+	"github.com/cosmos/cosmos-sdk/client/tx"
+	"github.com/spf13/cobra"
+
+	"github.com/desmos-labs/desmos/x/profiles/types"
+)
+
+// GetCmdCreateRelationship returns the command allowing to create a relationship
+func GetCmdCreateRelationship() *cobra.Command {
+	cmd := &cobra.Command{
+		Use:   "create-relationship [receiver] [subspace]",
+		Short: "Create a relationship with the given receiver address",
+		Args:  cobra.ExactArgs(2),
+		RunE: func(cmd *cobra.Command, args []string) error {
+			clientCtx, err := client.GetClientTxContext(cmd)
+			if err != nil {
+				return err
+			}
+
+			msg := types.NewMsgCreateRelationship(clientCtx.FromAddress.String(), args[0], args[1])
+			if err = msg.ValidateBasic(); err != nil {
+				return fmt.Errorf("message validation failed: %w", err)
+			}
+
+			return tx.GenerateOrBroadcastTxCLI(clientCtx, cmd.Flags(), msg)
+		},
+	}
+
+	flags.AddTxFlagsToCmd(cmd)
+
+	return cmd
+}
+
+// GetCmdDeleteRelationship returns the command allowing to delete a relationships
+func GetCmdDeleteRelationship() *cobra.Command {
+	cmd := &cobra.Command{
+		Use:   "delete-relationship [receiver] [subspace]",
+		Short: "Delete the relationship with the given user",
+		Args:  cobra.ExactArgs(2),
+		RunE: func(cmd *cobra.Command, args []string) error {
+			clientCtx, err := client.GetClientTxContext(cmd)
+			if err != nil {
+				return err
+			}
+
+			msg := types.NewMsgDeleteRelationship(clientCtx.FromAddress.String(), args[0], args[1])
+			if err = msg.ValidateBasic(); err != nil {
+				return fmt.Errorf("message validation failed: %w", err)
+			}
+
+			return tx.GenerateOrBroadcastTxCLI(clientCtx, cmd.Flags(), msg)
+		},
+	}
+
+	flags.AddTxFlagsToCmd(cmd)
+
+	return cmd
+}
+
+// GetCmdBlockUser returns the command allowing to block a user
+func GetCmdBlockUser() *cobra.Command {
+	cmd := &cobra.Command{
+		Use:   "block [address] [subspace] [[reason]]",
+		Short: "Block the user with the given address, optionally specifying the reason for the block",
+		Args:  cobra.RangeArgs(2, 3),
+		RunE: func(cmd *cobra.Command, args []string) error {
+			clientCtx, err := client.GetClientTxContext(cmd)
+			if err != nil {
+				return err
+			}
+
+			reason := ""
+			if len(args) == 3 {
+				reason = args[2]
+			}
+
+			msg := types.NewMsgBlockUser(clientCtx.FromAddress.String(), args[0], reason, args[1])
+			if err = msg.ValidateBasic(); err != nil {
+				return fmt.Errorf("message validation failed: %w", err)
+			}
+
+			return tx.GenerateOrBroadcastTxCLI(clientCtx, cmd.Flags(), msg)
+		},
+	}
+
+	flags.AddTxFlagsToCmd(cmd)
+
+	return cmd
+}
+
+// GetCmdUnblockUser returns the command allowing to unblock a user
+func GetCmdUnblockUser() *cobra.Command {
+	cmd := &cobra.Command{
+		Use:   "unblock [address] [subspace]",
+		Short: "Unblock the user with the given address",
+		Args:  cobra.ExactArgs(2),
+		RunE: func(cmd *cobra.Command, args []string) error {
+			clientCtx, err := client.GetClientTxContext(cmd)
+			if err != nil {
+				return err
+			}
+
+			msg := types.NewMsgUnblockUser(clientCtx.FromAddress.String(), args[0], args[1])
+			if err = msg.ValidateBasic(); err != nil {
+				return fmt.Errorf("message validation failed: %w", err)
+			}
+
+			return tx.GenerateOrBroadcastTxCLI(clientCtx, cmd.Flags(), msg)
+		},
+	}
+
+	flags.AddTxFlagsToCmd(cmd)
+
+	return cmd
+}
+
+// --------------------------------------------------------------------------------------------------------------------
+
+// GetCmdQueryUserRelationships returns the command allowing to query all the relationships of a specific user
+func GetCmdQueryUserRelationships() *cobra.Command {
+	cmd := &cobra.Command{
+		Use:   "relationships [address]",
+		Short: "Retrieve all the user's relationships",
+		Args:  cobra.ExactArgs(1),
+		RunE: func(cmd *cobra.Command, args []string) error {
+			clientCtx, err := client.GetClientQueryContext(cmd)
+			if err != nil {
+				return err
+			}
+			queryClient := types.NewQueryClient(clientCtx)
+
+			res, err := queryClient.UserRelationships(
+				context.Background(),
+				&types.QueryUserRelationshipsRequest{User: args[0]},
+			)
+			if err != nil {
+				return err
+			}
+
+			return clientCtx.PrintProto(res)
+		},
+	}
+
+	flags.AddQueryFlagsToCmd(cmd)
+
+	return cmd
+}
+
+// GetCmdQueryUserBlocks returns the command allowing to query all the blocks of a single user
+func GetCmdQueryUserBlocks() *cobra.Command {
+	cmd := &cobra.Command{
+		Use:   "blocklist [address]",
+		Short: "Retrieve the list of all the blocked users of the given address",
+		Args:  cobra.ExactArgs(1),
+		RunE: func(cmd *cobra.Command, args []string) error {
+			clientCtx, err := client.GetClientQueryContext(cmd)
+			if err != nil {
+				return err
+			}
+			queryClient := types.NewQueryClient(clientCtx)
+
+			res, err := queryClient.UserBlocks(
+				context.Background(),
+				&types.QueryUserBlocksRequest{User: args[0]})
+			if err != nil {
+				return err
+			}
+
+			return clientCtx.PrintProto(res)
+		},
+	}
+
+	flags.AddQueryFlagsToCmd(cmd)
+
+	return cmd
+}
