+++ x/profiles/client/cli/cli_test.go
@@ -1,9 +1,15 @@
 package cli_test
 
 import (
-	"fmt"
+	"encoding/hex"
+	"io/ioutil"
 	"testing"
 	"time"
+
+	cryptotypes "github.com/cosmos/cosmos-sdk/crypto/types"
+
+	"github.com/desmos-labs/desmos/app"
+	"github.com/desmos-labs/desmos/x/profiles/client/utils"
 
 	codectypes "github.com/cosmos/cosmos-sdk/codec/types"
 	"github.com/cosmos/cosmos-sdk/crypto/hd"
@@ -12,18 +18,23 @@ import (
 
 	authtypes "github.com/cosmos/cosmos-sdk/x/auth/types"
 
-	"github.com/cosmos/cosmos-sdk/client/flags"
-	clitestutil "github.com/cosmos/cosmos-sdk/testutil/cli"
 	"github.com/cosmos/cosmos-sdk/testutil/network"
 	sdk "github.com/cosmos/cosmos-sdk/types"
-	"github.com/golang/protobuf/proto"
 	"github.com/stretchr/testify/suite"
-	tmcli "github.com/tendermint/tendermint/libs/cli"
 
 	"github.com/desmos-labs/desmos/testutil"
-	"github.com/desmos-labs/desmos/x/profiles/client/cli"
 	"github.com/desmos-labs/desmos/x/profiles/types"
 )
+
+const (
+	srcKeyName = "src"
+	dstKeyName = "dst"
+)
+
+type Keys struct {
+	Src cryptotypes.PrivKey
+	Dst cryptotypes.PrivKey
+}
 
 type IntegrationTestSuite struct {
 	suite.Suite
@@ -31,6 +42,7 @@ type IntegrationTestSuite struct {
 	cfg         network.Config
 	network     *network.Network
 	keyBase     keyring.Keyring
+	keys        Keys
 	testProfile *types.Profile
 }
 
@@ -49,10 +61,13 @@ func (s *IntegrationTestSuite) SetupSuite() {
 	s.Require().NoError(cfg.Codec.UnmarshalJSON(genesisState[authtypes.ModuleName], &authData))
 
 	// Generate test keys
-	s.keyBase = generateMemoryKeybase()
-	srcKey, err := s.keyBase.Key("src")
+	s.setupKeyBase()
+
+	// Get keys
+	srcKey, err := s.keyBase.Key(srcKeyName)
 	s.Require().NoError(err)
-	destKey, err := s.keyBase.Key("dest")
+
+	destKey, err := s.keyBase.Key(dstKeyName)
 	s.Require().NoError(err)
 
 	// Store a profile account inside the auth genesis data
@@ -140,16 +155,10 @@ func (s *IntegrationTestSuite) SetupSuite() {
 	s.Require().NoError(err)
 }
 
-func (s *IntegrationTestSuite) TearDownSuite() {
-	s.T().Log("tearing down integration test suite")
-	s.network.Cleanup()
-}
-
-func generateMemoryKeybase() keyring.Keyring {
+func (s *IntegrationTestSuite) setupKeyBase() {
 	keyBase := keyring.NewInMemory()
-	keyringAlgos, _ := keyBase.SupportedAlgorithms()
-	algo, _ := keyring.NewSigningAlgoFromString("secp256k1", keyringAlgos)
-	hdPath := hd.CreateHDPath(0, 0, 0).String()
+	algo := hd.Secp256k1
+	hdPath := sdk.GetConfig().GetFullFundraiserPath()
 
 	srcEntropySeed, _ := bip39.NewEntropy(256)
 	destEntropySeed, _ := bip39.NewEntropy(256)
@@ -157,1151 +166,50 @@ func generateMemoryKeybase() keyring.Keyring {
 	srcMnemonic, _ := bip39.NewMnemonic(srcEntropySeed)
 	destMnemonic, _ := bip39.NewMnemonic(destEntropySeed)
 
-	keyBase.NewAccount("src", srcMnemonic, "", hdPath, algo)
-	keyBase.NewAccount("dest", destMnemonic, "", hdPath, algo)
-	return keyBase
-}
-
-// ___________________________________________________________________________________________________________________
-
-func (s *IntegrationTestSuite) TestCmdQueryProfile() {
-	val := s.network.Validators[0]
-
-	addr, err := sdk.AccAddressFromBech32("cosmos1ftkjv8njvkekk00ehwdfl5sst8zgdpenjfm4hs")
+	_, err := keyBase.NewAccount(srcKeyName, srcMnemonic, "", hdPath, algo)
 	s.Require().NoError(err)
 
-	profile, err := types.NewProfile(
-		"dtag",
-		"nickname",
-		"bio",
-		types.Pictures{},
-		time.Date(2020, 1, 1, 0, 0, 0, 0, time.UTC),
-		authtypes.NewBaseAccountWithAddress(addr),
+	_, err = keyBase.NewAccount(dstKeyName, destMnemonic, "", hdPath, algo)
+	s.Require().NoError(err)
+
+	s.keyBase = keyBase
+
+	derivedSrcPriv, err := algo.Derive()(srcMnemonic, "", hdPath)
+	s.Require().NoError(err)
+
+	derivedDstPriv, err := algo.Derive()(destMnemonic, "", hdPath)
+	s.Require().NoError(err)
+
+	s.keys = Keys{
+		Src: algo.Generate()(derivedSrcPriv),
+		Dst: algo.Generate()(derivedDstPriv),
+	}
+}
+
+func (s *IntegrationTestSuite) writeChainLinkJSONFile(filePath string) {
+	srcKey := s.keys.Src
+
+	addStr, err := sdk.Bech32ifyAddressBytes("cosmos", srcKey.PubKey().Address())
+	s.Require().NoError(err)
+
+	plainText := addStr
+	sigBz, err := srcKey.Sign([]byte(plainText))
+	s.Require().NoError(err)
+
+	jsonData := utils.NewChainLinkJSON(
+		types.NewBech32Address(addStr, "cosmos"),
+		types.NewProof(srcKey.PubKey(), hex.EncodeToString(sigBz), plainText),
+		types.NewChainConfig("cosmos"),
 	)
-	s.Require().NoError(err)
 
-	srcKey, err := s.keyBase.Key("src")
-	profile.ChainsLinks = []types.ChainLink{
-		types.NewChainLink(
-			types.NewBech32Address(srcKey.GetAddress().String(), "cosmos"),
-			types.NewProof(srcKey.GetPubKey(), "signature", "plain_text"),
-			types.NewChainConfig("cosmos"),
-			time.Date(2020, 1, 1, 0, 0, 0, 0, time.UTC),
-		),
-	}
+	params := app.MakeTestEncodingConfig()
+	jsonBz := params.Marshaler.MustMarshalJSON(&jsonData)
 
-	profileAny, err := codectypes.NewAnyWithValue(profile)
-	s.Require().NoError(err)
-
-	testCases := []struct {
-		name           string
-		args           []string
-		expectErr      bool
-		expectedOutput types.QueryProfileResponse
-	}{
-		{
-			name: "non existing profile",
-			args: []string{
-				s.network.Validators[1].Address.String(),
-				fmt.Sprintf("--%s=json", tmcli.OutputFlag),
-			},
-			expectErr: false,
-			expectedOutput: types.QueryProfileResponse{
-				Profile: nil,
-			},
-		},
-		{
-			name: "existing profile is returned properly",
-			args: []string{
-				"cosmos1ftkjv8njvkekk00ehwdfl5sst8zgdpenjfm4hs",
-				fmt.Sprintf("--%s=json", tmcli.OutputFlag),
-			},
-			expectErr: false,
-			expectedOutput: types.QueryProfileResponse{
-				Profile: profileAny,
-			},
-		},
-	}
-
-	for _, tc := range testCases {
-		tc := tc
-
-		s.Run(tc.name, func() {
-			cmd := cli.GetCmdQueryProfile()
-			clientCtx := val.ClientCtx
-			out, err := clitestutil.ExecTestCLICmd(clientCtx, cmd, tc.args)
-
-			if tc.expectErr {
-				s.Require().Error(err)
-			} else {
-				s.Require().NoError(err)
-
-				var response types.QueryProfileResponse
-				s.Require().NoError(clientCtx.JSONMarshaler.UnmarshalJSON(out.Bytes(), &response), out.String())
-				s.Require().True(tc.expectedOutput.Profile.Equal(response.Profile))
-			}
-		})
-	}
+	// Write the JSON to a temp file
+	s.Require().NoError(ioutil.WriteFile(filePath, jsonBz, 0666))
 }
 
-func (s *IntegrationTestSuite) TestCmdQueryDTagRequests() {
-	val := s.network.Validators[0]
-
-	testCases := []struct {
-		name           string
-		args           []string
-		expectErr      bool
-		expectedOutput types.QueryDTagTransfersResponse
-	}{
-		{
-			name: "empty slice is returned properly",
-			args: []string{
-				"cosmos1nqwf7chwfywdw2379sxmwlcgcfvvy86t6mpunz",
-				fmt.Sprintf("--%s=json", tmcli.OutputFlag),
-			},
-			expectErr: false,
-			expectedOutput: types.QueryDTagTransfersResponse{
-				Requests: []types.DTagTransferRequest{},
-			},
-		},
-		{
-			name: "existing requests are returned properly",
-			args: []string{
-				"cosmos1ftkjv8njvkekk00ehwdfl5sst8zgdpenjfm4hs",
-				fmt.Sprintf("--%s=json", tmcli.OutputFlag),
-			},
-			expectErr: false,
-			expectedOutput: types.QueryDTagTransfersResponse{
-				Requests: []types.DTagTransferRequest{
-					types.NewDTagTransferRequest(
-						"dtag",
-						"cosmos122u6u9gpdr2rp552fkkvlgyecjlmtqhkascl5a",
-						"cosmos1ftkjv8njvkekk00ehwdfl5sst8zgdpenjfm4hs",
-					),
-				},
-			},
-		},
-	}
-
-	for _, tc := range testCases {
-		tc := tc
-
-		s.Run(tc.name, func() {
-			cmd := cli.GetCmdQueryDTagRequests()
-			clientCtx := val.ClientCtx
-			out, err := clitestutil.ExecTestCLICmd(clientCtx, cmd, tc.args)
-
-			if tc.expectErr {
-				s.Require().Error(err)
-			} else {
-				s.Require().NoError(err)
-
-				var response types.QueryDTagTransfersResponse
-				s.Require().NoError(clientCtx.JSONMarshaler.UnmarshalJSON(out.Bytes(), &response), out.String())
-				s.Require().Equal(tc.expectedOutput, response)
-			}
-		})
-	}
-}
-
-func (s *IntegrationTestSuite) TestCmdQueryParams() {
-	val := s.network.Validators[0]
-
-	testCases := []struct {
-		name           string
-		args           []string
-		expectErr      bool
-		expectedOutput types.QueryParamsResponse
-	}{
-		{
-			name:      "existing params are returned properly",
-			args:      []string{fmt.Sprintf("--%s=json", tmcli.OutputFlag)},
-			expectErr: false,
-			expectedOutput: types.QueryParamsResponse{
-				Params: types.DefaultParams(),
-			},
-		},
-	}
-
-	for _, tc := range testCases {
-		tc := tc
-
-		s.Run(tc.name, func() {
-			cmd := cli.GetCmdQueryParams()
-			clientCtx := val.ClientCtx
-			out, err := clitestutil.ExecTestCLICmd(clientCtx, cmd, tc.args)
-
-			if tc.expectErr {
-				s.Require().Error(err)
-			} else {
-				s.Require().NoError(err)
-
-				var response types.QueryParamsResponse
-				s.Require().NoError(clientCtx.JSONMarshaler.UnmarshalJSON(out.Bytes(), &response), out.String())
-				s.Require().Equal(tc.expectedOutput, response)
-			}
-		})
-	}
-}
-
-func (s *IntegrationTestSuite) TestCmdQueryUserRelationships() {
-	val := s.network.Validators[0]
-
-	testCases := []struct {
-		name           string
-		args           []string
-		expectErr      bool
-		expectedOutput types.QueryUserRelationshipsResponse
-	}{
-		{
-			name: "empty array is returned properly",
-			args: []string{
-				s.network.Validators[1].Address.String(),
-				fmt.Sprintf("--%s=json", tmcli.OutputFlag),
-			},
-			expectErr: false,
-			expectedOutput: types.QueryUserRelationshipsResponse{
-				User:          s.network.Validators[1].Address.String(),
-				Relationships: []types.Relationship{},
-			},
-		},
-		{
-			name: "existing relationship is returned properly",
-			args: []string{
-				"cosmos122u6u9gpdr2rp552fkkvlgyecjlmtqhkascl5a",
-				fmt.Sprintf("--%s=json", tmcli.OutputFlag),
-			},
-			expectErr: false,
-			expectedOutput: types.QueryUserRelationshipsResponse{
-				User: "cosmos122u6u9gpdr2rp552fkkvlgyecjlmtqhkascl5a",
-				Relationships: []types.Relationship{
-					types.NewRelationship(
-						"cosmos122u6u9gpdr2rp552fkkvlgyecjlmtqhkascl5a",
-						"cosmos1ftkjv8njvkekk00ehwdfl5sst8zgdpenjfm4hs",
-						"60303ae22b998861bce3b28f33eec1be758a213c86c93c076dbe9f558c11c752",
-					),
-				},
-			},
-		},
-	}
-
-	for _, tc := range testCases {
-		tc := tc
-
-		s.Run(tc.name, func() {
-			cmd := cli.GetCmdQueryUserRelationships()
-			clientCtx := val.ClientCtx
-			out, err := clitestutil.ExecTestCLICmd(clientCtx, cmd, tc.args)
-
-			if tc.expectErr {
-				s.Require().Error(err)
-			} else {
-				s.Require().NoError(err)
-
-				var response types.QueryUserRelationshipsResponse
-				s.Require().NoError(clientCtx.JSONMarshaler.UnmarshalJSON(out.Bytes(), &response), out.String())
-				s.Require().Equal(tc.expectedOutput, response)
-			}
-		})
-	}
-}
-
-func (s *IntegrationTestSuite) TestCmdQueryUserBlocks() {
-	val := s.network.Validators[0]
-
-	testCases := []struct {
-		name           string
-		args           []string
-		expectErr      bool
-		expectedOutput types.QueryUserBlocksResponse
-	}{
-		{
-			name: "empty slice is returned properly",
-			args: []string{
-				s.network.Validators[1].Address.String(),
-				fmt.Sprintf("--%s=json", tmcli.OutputFlag),
-			},
-			expectErr: false,
-			expectedOutput: types.QueryUserBlocksResponse{
-				Blocks: []types.UserBlock{},
-			},
-		},
-		{
-			name: "existing user blocks are returned properly",
-			args: []string{
-				"cosmos1azqm9kmyxunkx2yt332hmnr8sa3lclhjlg9w5k",
-				fmt.Sprintf("--%s=json", tmcli.OutputFlag),
-			},
-			expectErr: false,
-			expectedOutput: types.QueryUserBlocksResponse{
-				Blocks: []types.UserBlock{
-					types.NewUserBlock(
-						"cosmos1azqm9kmyxunkx2yt332hmnr8sa3lclhjlg9w5k",
-						"cosmos1zs70glquczqgt83g03jnvcqppu4jjj8yjxwlvh",
-						"Test block",
-						"9f86d081884c7d659a2feaa0c55ad015a3bf4f1b2b0b822cd15d6c15b0f00a08",
-					),
-				},
-			},
-		},
-	}
-
-	for _, tc := range testCases {
-		tc := tc
-
-		s.Run(tc.name, func() {
-			cmd := cli.GetCmdQueryUserBlocks()
-			clientCtx := val.ClientCtx
-			out, err := clitestutil.ExecTestCLICmd(clientCtx, cmd, tc.args)
-
-			if tc.expectErr {
-				s.Require().Error(err)
-			} else {
-				s.Require().NoError(err)
-
-				var response types.QueryUserBlocksResponse
-				s.Require().NoError(clientCtx.JSONMarshaler.UnmarshalJSON(out.Bytes(), &response), out.String())
-				s.Require().Equal(tc.expectedOutput, response)
-			}
-		})
-	}
-}
-
-func (s *IntegrationTestSuite) TestCmdQueryProfileByChainLink() {
-	val := s.network.Validators[0]
-
-	addr, err := sdk.AccAddressFromBech32("cosmos1ftkjv8njvkekk00ehwdfl5sst8zgdpenjfm4hs")
-	s.Require().NoError(err)
-
-	profile, err := types.NewProfile(
-		"dtag",
-		"nickname",
-		"bio",
-		types.Pictures{},
-		time.Date(2020, 1, 1, 0, 0, 0, 0, time.UTC),
-		authtypes.NewBaseAccountWithAddress(addr),
-	)
-	s.Require().NoError(err)
-
-	srcKey, err := s.keyBase.Key("src")
-	profile.ChainsLinks = []types.ChainLink{
-		types.NewChainLink(
-			types.NewBech32Address(srcKey.GetAddress().String(), "cosmos"),
-			types.NewProof(srcKey.GetPubKey(), "signature", "plain_text"),
-			types.NewChainConfig("cosmos"),
-			time.Date(2020, 1, 1, 0, 0, 0, 0, time.UTC),
-		),
-	}
-
-	profileAny, err := codectypes.NewAnyWithValue(profile)
-	s.Require().NoError(err)
-
-	testCases := []struct {
-		name           string
-		args           []string
-		expectErr      bool
-		expectedOutput types.QueryProfileByChainLinkResponse
-	}{
-		{
-			name: "no link associated to profile",
-			args: []string{
-				"cosmos",
-				s.network.Validators[1].Address.String(),
-				fmt.Sprintf("--%s=json", tmcli.OutputFlag),
-			},
-			expectErr: true,
-			expectedOutput: types.QueryProfileByChainLinkResponse{
-				Profile: nil,
-			},
-		},
-		{
-			name: "existing profile is returned properly",
-			args: []string{
-				"cosmos",
-				srcKey.GetAddress().String(),
-				fmt.Sprintf("--%s=json", tmcli.OutputFlag),
-			},
-			expectErr: false,
-			expectedOutput: types.QueryProfileByChainLinkResponse{
-				Profile: profileAny,
-			},
-		},
-	}
-
-	for _, tc := range testCases {
-		tc := tc
-
-		s.Run(tc.name, func() {
-			cmd := cli.GetCmdQueryProfileByChainLink()
-			clientCtx := val.ClientCtx
-			out, err := clitestutil.ExecTestCLICmd(clientCtx, cmd, tc.args)
-
-			if tc.expectErr {
-				s.Require().Error(err)
-			} else {
-				s.Require().NoError(err)
-
-				var response types.QueryProfileResponse
-				s.Require().NoError(clientCtx.JSONMarshaler.UnmarshalJSON(out.Bytes(), &response), out.String())
-				s.Require().True(tc.expectedOutput.Profile.Equal(response.Profile))
-			}
-		})
-	}
-}
-
-// ___________________________________________________________________________________________________________________
-
-func (s *IntegrationTestSuite) TestCmdSaveProfile() {
-	val := s.network.Validators[0]
-
-	testCases := []struct {
-		name     string
-		args     []string
-		expErr   bool
-		respType proto.Message
-	}{
-		{
-			name: "invalid dtag returns error",
-			args: []string{
-				"",
-				fmt.Sprintf("--%s=%s", flags.FlagFrom, val.Address.String()),
-			},
-			expErr: true,
-		},
-		{
-			name: "invalid creator returns error",
-			args: []string{
-				"dtag",
-				fmt.Sprintf("--%s=%s", flags.FlagFrom, ""),
-			},
-			expErr: true,
-		},
-		{
-			name: "correct data returns no error",
-			args: []string{
-				"dtag",
-				fmt.Sprintf("--%s=%s", flags.FlagFrom, val.Address.String()),
-				fmt.Sprintf("--%s=true", flags.FlagSkipConfirmation),
-				fmt.Sprintf("--%s=%s", flags.FlagBroadcastMode, flags.BroadcastBlock),
-				fmt.Sprintf("--%s=%s", flags.FlagFees, sdk.NewCoins(sdk.NewCoin(s.cfg.BondDenom, sdk.NewInt(10))).String()),
-			},
-			expErr:   false,
-			respType: &sdk.TxResponse{},
-		},
-	}
-
-	for _, tc := range testCases {
-		tc := tc
-
-		s.Run(tc.name, func() {
-			cmd := cli.GetCmdSaveProfile()
-			clientCtx := val.ClientCtx
-
-			out, err := clitestutil.ExecTestCLICmd(clientCtx, cmd, tc.args)
-			if tc.expErr {
-				s.Require().Error(err)
-			} else {
-				s.Require().NoError(err)
-				s.Require().NoError(clientCtx.JSONMarshaler.UnmarshalJSON(out.Bytes(), tc.respType), out.String())
-			}
-		})
-	}
-}
-
-func (s *IntegrationTestSuite) TestCmdDeleteProfile() {
-	val := s.network.Validators[0]
-
-	testCases := []struct {
-		name     string
-		args     []string
-		expErr   bool
-		respType proto.Message
-	}{
-		{
-			name:   "invalid user returns error",
-			args:   []string{fmt.Sprintf("--%s=%s", flags.FlagFrom, "")},
-			expErr: true,
-		},
-		{
-			name: "correct data returns no error",
-			args: []string{
-				fmt.Sprintf("--%s=%s", flags.FlagFrom, val.Address.String()),
-				fmt.Sprintf("--%s=true", flags.FlagSkipConfirmation),
-				fmt.Sprintf("--%s=%s", flags.FlagBroadcastMode, flags.BroadcastBlock),
-				fmt.Sprintf("--%s=%s", flags.FlagFees, sdk.NewCoins(sdk.NewCoin(s.cfg.BondDenom, sdk.NewInt(10))).String()),
-			},
-			expErr:   false,
-			respType: &sdk.TxResponse{},
-		},
-	}
-
-	for _, tc := range testCases {
-		tc := tc
-
-		s.Run(tc.name, func() {
-			cmd := cli.GetCmdDeleteProfile()
-			clientCtx := val.ClientCtx
-
-			out, err := clitestutil.ExecTestCLICmd(clientCtx, cmd, tc.args)
-			if tc.expErr {
-				s.Require().Error(err)
-			} else {
-				s.Require().NoError(err)
-				s.Require().NoError(clientCtx.JSONMarshaler.UnmarshalJSON(out.Bytes(), tc.respType), out.String())
-			}
-		})
-	}
-}
-
-func (s *IntegrationTestSuite) TestCmdRequestDTagTransfer() {
-	val := s.network.Validators[0]
-
-	testCases := []struct {
-		name     string
-		args     []string
-		expErr   bool
-		respType proto.Message
-	}{
-		{
-			name: "invalid user address returns error",
-			args: []string{
-				"",
-				fmt.Sprintf("--%s=%s", flags.FlagFrom, val.Address.String()),
-			},
-			expErr: true,
-		},
-		{
-			name: "same user returns error",
-			args: []string{
-				val.Address.String(),
-				fmt.Sprintf("--%s=%s", flags.FlagFrom, val.Address.String()),
-			},
-			expErr: true,
-		},
-		{
-			name: "correct data returns no error",
-			args: []string{
-				s.network.Validators[1].Address.String(),
-				fmt.Sprintf("--%s=%s", flags.FlagFrom, val.Address.String()),
-				fmt.Sprintf("--%s=true", flags.FlagSkipConfirmation),
-				fmt.Sprintf("--%s=%s", flags.FlagBroadcastMode, flags.BroadcastBlock),
-				fmt.Sprintf("--%s=%s", flags.FlagFees, sdk.NewCoins(sdk.NewCoin(s.cfg.BondDenom, sdk.NewInt(10))).String()),
-			},
-			expErr:   false,
-			respType: &sdk.TxResponse{},
-		},
-	}
-
-	for _, tc := range testCases {
-		tc := tc
-
-		s.Run(tc.name, func() {
-			cmd := cli.GetCmdRequestDTagTransfer()
-			clientCtx := val.ClientCtx
-
-			out, err := clitestutil.ExecTestCLICmd(clientCtx, cmd, tc.args)
-			if tc.expErr {
-				s.Require().Error(err)
-			} else {
-				s.Require().NoError(err)
-				s.Require().NoError(clientCtx.JSONMarshaler.UnmarshalJSON(out.Bytes(), tc.respType), out.String())
-			}
-		})
-	}
-}
-
-func (s *IntegrationTestSuite) TestCmdCancelDTagTransfer() {
-	val := s.network.Validators[0]
-
-	testCases := []struct {
-		name     string
-		args     []string
-		expErr   bool
-		respType proto.Message
-	}{
-		{
-			name: "invalid recipient returns error",
-			args: []string{
-				"",
-				fmt.Sprintf("--%s=%s", flags.FlagFrom, val.Address.String()),
-			},
-			expErr: true,
-		},
-		{
-			name: "same user and recipient returns error",
-			args: []string{
-				val.Address.String(),
-				fmt.Sprintf("--%s=%s", flags.FlagFrom, val.Address.String()),
-			},
-			expErr: true,
-		},
-		{
-			name: "correct data returns no error",
-			args: []string{
-				s.network.Validators[1].Address.String(),
-				fmt.Sprintf("--%s=%s", flags.FlagFrom, val.Address.String()),
-				fmt.Sprintf("--%s=true", flags.FlagSkipConfirmation),
-				fmt.Sprintf("--%s=%s", flags.FlagBroadcastMode, flags.BroadcastBlock),
-				fmt.Sprintf("--%s=%s", flags.FlagFees, sdk.NewCoins(sdk.NewCoin(s.cfg.BondDenom, sdk.NewInt(10))).String()),
-			},
-			expErr:   false,
-			respType: &sdk.TxResponse{},
-		},
-	}
-
-	for _, tc := range testCases {
-		tc := tc
-
-		s.Run(tc.name, func() {
-			cmd := cli.GetCmdCancelDTagTransfer()
-			clientCtx := val.ClientCtx
-
-			out, err := clitestutil.ExecTestCLICmd(clientCtx, cmd, tc.args)
-			if tc.expErr {
-				s.Require().Error(err)
-			} else {
-				s.Require().NoError(err)
-				s.Require().NoError(clientCtx.JSONMarshaler.UnmarshalJSON(out.Bytes(), tc.respType), out.String())
-			}
-		})
-	}
-}
-
-func (s *IntegrationTestSuite) TestCmdAcceptDTagTransfer() {
-	val := s.network.Validators[0]
-
-	testCases := []struct {
-		name     string
-		args     []string
-		expErr   bool
-		respType proto.Message
-	}{
-		{
-			name: "invalid new dtag returns error",
-			args: []string{
-				"",
-				s.network.Validators[1].Address.String(),
-				fmt.Sprintf("--%s=%s", flags.FlagFrom, val.Address.String()),
-			},
-			expErr: true,
-		},
-		{
-			name: "invalid request sender returns error",
-			args: []string{
-				"dtag",
-				"",
-				fmt.Sprintf("--%s=%s", flags.FlagFrom, val.Address.String()),
-			},
-			expErr: true,
-		},
-		{
-			name: "correct data returns no error",
-			args: []string{
-				"dtag",
-				s.network.Validators[1].Address.String(),
-				fmt.Sprintf("--%s=%s", flags.FlagFrom, val.Address.String()),
-				fmt.Sprintf("--%s=true", flags.FlagSkipConfirmation),
-				fmt.Sprintf("--%s=%s", flags.FlagBroadcastMode, flags.BroadcastBlock),
-				fmt.Sprintf("--%s=%s", flags.FlagFees, sdk.NewCoins(sdk.NewCoin(s.cfg.BondDenom, sdk.NewInt(10))).String()),
-			},
-			expErr:   false,
-			respType: &sdk.TxResponse{},
-		},
-	}
-
-	for _, tc := range testCases {
-		tc := tc
-
-		s.Run(tc.name, func() {
-			cmd := cli.GetCmdAcceptDTagTransfer()
-			clientCtx := val.ClientCtx
-
-			out, err := clitestutil.ExecTestCLICmd(clientCtx, cmd, tc.args)
-			if tc.expErr {
-				s.Require().Error(err)
-			} else {
-				s.Require().NoError(err)
-				s.Require().NoError(clientCtx.JSONMarshaler.UnmarshalJSON(out.Bytes(), tc.respType), out.String())
-			}
-		})
-	}
-}
-
-func (s *IntegrationTestSuite) TestCmdRefuseDTagTransfer() {
-	val := s.network.Validators[0]
-
-	testCases := []struct {
-		name     string
-		args     []string
-		expErr   bool
-		respType proto.Message
-	}{
-		{
-			name: "invalid sender returns error",
-			args: []string{
-				"",
-				fmt.Sprintf("--%s=%s", flags.FlagFrom, val.Address.String()),
-			},
-			expErr: true,
-		},
-		{
-			name: "same user and sender returns error",
-			args: []string{
-				val.Address.String(),
-				fmt.Sprintf("--%s=%s", flags.FlagFrom, val.Address.String()),
-			},
-			expErr: true,
-		},
-		{
-			name: "correct data returns no error",
-			args: []string{
-				s.network.Validators[1].Address.String(),
-				fmt.Sprintf("--%s=%s", flags.FlagFrom, val.Address.String()),
-				fmt.Sprintf("--%s=true", flags.FlagSkipConfirmation),
-				fmt.Sprintf("--%s=%s", flags.FlagBroadcastMode, flags.BroadcastBlock),
-				fmt.Sprintf("--%s=%s", flags.FlagFees, sdk.NewCoins(sdk.NewCoin(s.cfg.BondDenom, sdk.NewInt(10))).String()),
-			},
-			expErr:   false,
-			respType: &sdk.TxResponse{},
-		},
-	}
-
-	for _, tc := range testCases {
-		tc := tc
-
-		s.Run(tc.name, func() {
-			cmd := cli.GetCmdRefuseDTagTransfer()
-			clientCtx := val.ClientCtx
-
-			out, err := clitestutil.ExecTestCLICmd(clientCtx, cmd, tc.args)
-			if tc.expErr {
-				s.Require().Error(err)
-			} else {
-				s.Require().NoError(err)
-				s.Require().NoError(clientCtx.JSONMarshaler.UnmarshalJSON(out.Bytes(), tc.respType), out.String())
-			}
-		})
-	}
-}
-
-func (s *IntegrationTestSuite) TestCmdCreateRelationship() {
-	val := s.network.Validators[0]
-
-	testCases := []struct {
-		name     string
-		args     []string
-		expErr   bool
-		respType proto.Message
-	}{
-		{
-			name: "invalid subspace returns error",
-			args: []string{
-				val.Address.String(),
-				"subspace",
-				fmt.Sprintf("--%s=%s", flags.FlagFrom, val.Address.String()),
-			},
-			expErr: true,
-		},
-		{
-			name: "invalid blocked user returns error",
-			args: []string{
-				"address",
-				"9f86d081884c7d659a2feaa0c55ad015a3bf4f1b2b0b822cd15d6c15b0f00a08",
-				fmt.Sprintf("--%s=%s", flags.FlagFrom, val.Address.String()),
-			},
-			expErr: true,
-		},
-		{
-			name: "same user and counterparty returns error",
-			args: []string{
-				val.Address.String(),
-				"9f86d081884c7d659a2feaa0c55ad015a3bf4f1b2b0b822cd15d6c15b0f00a08",
-				fmt.Sprintf("--%s=%s", flags.FlagFrom, val.Address.String()),
-			},
-			expErr: true,
-		},
-		{
-			name: "valid parameters works properly",
-			args: []string{
-				s.network.Validators[1].Address.String(),
-				"9f86d081884c7d659a2feaa0c55ad015a3bf4f1b2b0b822cd15d6c15b0f00a08",
-				fmt.Sprintf("--%s=%s", flags.FlagFrom, val.Address.String()),
-				fmt.Sprintf("--%s=true", flags.FlagSkipConfirmation),
-				fmt.Sprintf("--%s=%s", flags.FlagBroadcastMode, flags.BroadcastBlock),
-				fmt.Sprintf("--%s=%s", flags.FlagFees, sdk.NewCoins(sdk.NewCoin(s.cfg.BondDenom, sdk.NewInt(10))).String()),
-			},
-			expErr:   false,
-			respType: &sdk.TxResponse{},
-		},
-	}
-
-	for _, tc := range testCases {
-		tc := tc
-
-		s.Run(tc.name, func() {
-			cmd := cli.GetCmdCreateRelationship()
-			clientCtx := val.ClientCtx
-
-			out, err := clitestutil.ExecTestCLICmd(clientCtx, cmd, tc.args)
-			if tc.expErr {
-				s.Require().Error(err)
-			} else {
-				s.Require().NoError(err)
-				s.Require().NoError(clientCtx.JSONMarshaler.UnmarshalJSON(out.Bytes(), tc.respType), out.String())
-			}
-		})
-	}
-}
-
-func (s *IntegrationTestSuite) TestCmdDeleteRelationship() {
-	val := s.network.Validators[0]
-
-	testCases := []struct {
-		name     string
-		args     []string
-		expErr   bool
-		respType proto.Message
-	}{
-		{
-			name: "invalid receiver returns error",
-			args: []string{
-				"receiver",
-				"9f86d081884c7d659a2feaa0c55ad015a3bf4f1b2b0b822cd15d6c15b0f00a08",
-				fmt.Sprintf("--%s=%s", flags.FlagFrom, val.Address.String()),
-			},
-			expErr: true,
-		},
-		{
-			name: "invalid subspace returns error",
-			args: []string{
-				s.network.Validators[1].Address.String(),
-				"subspace",
-				fmt.Sprintf("--%s=%s", flags.FlagFrom, val.Address.String()),
-			},
-			expErr: true,
-		},
-		{
-			name: "same user and counterparty returns error",
-			args: []string{
-				val.Address.String(),
-				"9f86d081884c7d659a2feaa0c55ad015a3bf4f1b2b0b822cd15d6c15b0f00a08",
-				fmt.Sprintf("--%s=%s", flags.FlagFrom, val.Address.String()),
-			},
-			expErr: true,
-		},
-		{
-			name: "valid request executes properly",
-			args: []string{
-				s.network.Validators[1].Address.String(),
-				"9f86d081884c7d659a2feaa0c55ad015a3bf4f1b2b0b822cd15d6c15b0f00a08",
-				fmt.Sprintf("--%s=%s", flags.FlagFrom, val.Address.String()),
-				fmt.Sprintf("--%s=true", flags.FlagSkipConfirmation),
-				fmt.Sprintf("--%s=%s", flags.FlagBroadcastMode, flags.BroadcastBlock),
-				fmt.Sprintf("--%s=%s", flags.FlagFees, sdk.NewCoins(sdk.NewCoin(s.cfg.BondDenom, sdk.NewInt(10))).String()),
-			},
-			expErr:   false,
-			respType: &sdk.TxResponse{},
-		},
-	}
-
-	for _, tc := range testCases {
-		tc := tc
-
-		s.Run(tc.name, func() {
-			cmd := cli.GetCmdDeleteRelationship()
-			clientCtx := val.ClientCtx
-
-			out, err := clitestutil.ExecTestCLICmd(clientCtx, cmd, tc.args)
-			if tc.expErr {
-				s.Require().Error(err)
-			} else {
-				s.Require().NoError(err)
-				s.Require().NoError(clientCtx.JSONMarshaler.UnmarshalJSON(out.Bytes(), tc.respType), out.String())
-			}
-		})
-	}
-}
-
-func (s *IntegrationTestSuite) TestCmdBlockUser() {
-	val := s.network.Validators[0]
-
-	testCases := []struct {
-		name     string
-		args     []string
-		expErr   bool
-		respType proto.Message
-	}{
-		{
-			name: "invalid blocked address returns error",
-			args: []string{
-				"address",
-				"9f86d081884c7d659a2feaa0c55ad015a3bf4f1b2b0b822cd15d6c15b0f00a08",
-				fmt.Sprintf("--%s=%s", flags.FlagFrom, val.Address.String()),
-			},
-			expErr: true,
-		},
-		{
-			name: "invalid subspace returns error",
-			args: []string{
-				s.network.Validators[1].Address.String(),
-				"subspace",
-				fmt.Sprintf("--%s=%s", flags.FlagFrom, val.Address.String()),
-			},
-			expErr: true,
-		},
-		{
-			name: "same blocker and blocked returns error",
-			args: []string{
-				val.Address.String(),
-				"9f86d081884c7d659a2feaa0c55ad015a3bf4f1b2b0b822cd15d6c15b0f00a08",
-				fmt.Sprintf("--%s=%s", flags.FlagFrom, val.Address.String()),
-			},
-			expErr: true,
-		},
-		{
-			name: "valid request works properly without reason",
-			args: []string{
-				s.network.Validators[1].Address.String(),
-				"9f86d081884c7d659a2feaa0c55ad015a3bf4f1b2b0b822cd15d6c15b0f00a08",
-				fmt.Sprintf("--%s=%s", flags.FlagFrom, val.Address.String()),
-				fmt.Sprintf("--%s=true", flags.FlagSkipConfirmation),
-				fmt.Sprintf("--%s=%s", flags.FlagBroadcastMode, flags.BroadcastBlock),
-				fmt.Sprintf("--%s=%s", flags.FlagFees, sdk.NewCoins(sdk.NewCoin(s.cfg.BondDenom, sdk.NewInt(10))).String()),
-			},
-			expErr:   false,
-			respType: &sdk.TxResponse{},
-		},
-		{
-			name: "valid request works properly with reason",
-			args: []string{
-				s.network.Validators[1].Address.String(),
-				"9f86d081884c7d659a2feaa0c55ad015a3bf4f1b2b0b822cd15d6c15b0f00a08",
-				"Blocking reason",
-				fmt.Sprintf("--%s=%s", flags.FlagFrom, val.Address.String()),
-				fmt.Sprintf("--%s=true", flags.FlagSkipConfirmation),
-				fmt.Sprintf("--%s=%s", flags.FlagBroadcastMode, flags.BroadcastBlock),
-				fmt.Sprintf("--%s=%s", flags.FlagFees, sdk.NewCoins(sdk.NewCoin(s.cfg.BondDenom, sdk.NewInt(10))).String()),
-			},
-			expErr:   false,
-			respType: &sdk.TxResponse{},
-		},
-	}
-
-	for _, tc := range testCases {
-		tc := tc
-
-		s.Run(tc.name, func() {
-			cmd := cli.GetCmdBlockUser()
-			clientCtx := val.ClientCtx
-
-			out, err := clitestutil.ExecTestCLICmd(clientCtx, cmd, tc.args)
-			if tc.expErr {
-				s.Require().Error(err)
-			} else {
-				s.Require().NoError(err)
-				s.Require().NoError(clientCtx.JSONMarshaler.UnmarshalJSON(out.Bytes(), tc.respType), out.String())
-			}
-		})
-	}
-}
-
-func (s *IntegrationTestSuite) TestCmdUnblockUser() {
-	val := s.network.Validators[0]
-
-	testCases := []struct {
-		name     string
-		args     []string
-		expErr   bool
-		respType proto.Message
-	}{
-		{
-			name: "invalid blocked address returns error",
-			args: []string{
-				"address",
-				"9f86d081884c7d659a2feaa0c55ad015a3bf4f1b2b0b822cd15d6c15b0f00a08",
-				fmt.Sprintf("--%s=%s", flags.FlagFrom, val.Address.String()),
-			},
-			expErr: true,
-		},
-		{
-			name: "invalid subspace returns error",
-			args: []string{
-				s.network.Validators[1].Address.String(),
-				"subspace",
-				fmt.Sprintf("--%s=%s", flags.FlagFrom, val.Address.String()),
-			},
-			expErr: true,
-		},
-		{
-			name: "same blocker and blocked returns error",
-			args: []string{
-				val.Address.String(),
-				"9f86d081884c7d659a2feaa0c55ad015a3bf4f1b2b0b822cd15d6c15b0f00a08",
-				fmt.Sprintf("--%s=%s", flags.FlagFrom, val.Address.String()),
-			},
-			expErr: true,
-		},
-		{
-			name: "valid request works properly without reason",
-			args: []string{
-				s.network.Validators[1].Address.String(),
-				"9f86d081884c7d659a2feaa0c55ad015a3bf4f1b2b0b822cd15d6c15b0f00a08",
-				fmt.Sprintf("--%s=%s", flags.FlagFrom, val.Address.String()),
-				fmt.Sprintf("--%s=true", flags.FlagSkipConfirmation),
-				fmt.Sprintf("--%s=%s", flags.FlagBroadcastMode, flags.BroadcastBlock),
-				fmt.Sprintf("--%s=%s", flags.FlagFees, sdk.NewCoins(sdk.NewCoin(s.cfg.BondDenom, sdk.NewInt(10))).String()),
-			},
-			expErr:   false,
-			respType: &sdk.TxResponse{},
-		},
-		{
-			name: "valid request works properly with reason",
-			args: []string{
-				s.network.Validators[1].Address.String(),
-				"9f86d081884c7d659a2feaa0c55ad015a3bf4f1b2b0b822cd15d6c15b0f00a08",
-				fmt.Sprintf("--%s=%s", flags.FlagFrom, val.Address.String()),
-				fmt.Sprintf("--%s=true", flags.FlagSkipConfirmation),
-				fmt.Sprintf("--%s=%s", flags.FlagBroadcastMode, flags.BroadcastBlock),
-				fmt.Sprintf("--%s=%s", flags.FlagFees, sdk.NewCoins(sdk.NewCoin(s.cfg.BondDenom, sdk.NewInt(10))).String()),
-			},
-			expErr:   false,
-			respType: &sdk.TxResponse{},
-		},
-	}
-
-	for _, tc := range testCases {
-		tc := tc
-
-		s.Run(tc.name, func() {
-			cmd := cli.GetCmdUnblockUser()
-			clientCtx := val.ClientCtx
-
-			out, err := clitestutil.ExecTestCLICmd(clientCtx, cmd, tc.args)
-			if tc.expErr {
-				s.Require().Error(err)
-			} else {
-				s.Require().NoError(err)
-				s.Require().NoError(clientCtx.JSONMarshaler.UnmarshalJSON(out.Bytes(), tc.respType), out.String())
-			}
-		})
-	}
-}
-
-func (s *IntegrationTestSuite) TestCmdLinkChainAccount() {
-	cliCtx := s.network.Validators[0].ClientCtx
-	cliCtx.Keyring = s.keyBase
-	testCases := []struct {
-		name     string
-		args     []string
-		expErr   bool
-		respType proto.Message
-	}{
-		{
-			name: "could not get destination key returns error",
-			args: []string{
-				"src",
-				fmt.Sprintf("--%s=%s", flags.FlagFrom, ""),
-			},
-			expErr:   true,
-			respType: &sdk.TxResponse{},
-		},
-		{
-			name: "could not get source key returns error",
-			args: []string{
-				"wrong",
-				fmt.Sprintf("--%s=%s", flags.FlagFrom, "dest"),
-			},
-			expErr:   true,
-			respType: &sdk.TxResponse{},
-		},
-		{
-			name: "valid request works properly",
-			args: []string{
-				"src",
-				fmt.Sprintf("--%s=%s", flags.FlagFrom, "dest"),
-				fmt.Sprintf("--%s=true", flags.FlagSkipConfirmation),
-				fmt.Sprintf("--%s=%s", flags.FlagBroadcastMode, flags.BroadcastBlock),
-				fmt.Sprintf("--%s=%s", flags.FlagFees, sdk.NewCoins(sdk.NewCoin(s.cfg.BondDenom, sdk.NewInt(10))).String()),
-			},
-			expErr:   false,
-			respType: &sdk.TxResponse{},
-		},
-	}
-
-	for _, tc := range testCases {
-		tc := tc
-
-		s.Run(tc.name, func() {
-			cmd := cli.GetCmdLinkChainAccount()
-			out, err := clitestutil.ExecTestCLICmd(cliCtx, cmd, tc.args)
-
-			if tc.expErr {
-				s.Require().Error(err)
-			} else {
-				s.Require().NoError(err)
-				s.Require().NoError(cliCtx.JSONMarshaler.UnmarshalJSON(out.Bytes(), tc.respType), out.String())
-			}
-		})
-	}
-}
-
-func (s *IntegrationTestSuite) TestCmdUnlinkChainAccount() {
-	cliCtx := s.network.Validators[0].ClientCtx
-	cliCtx.Keyring = s.keyBase
-	src, err := s.keyBase.Key("src")
-	s.Require().NoError(err)
-	testCases := []struct {
-		name     string
-		args     []string
-		expErr   bool
-		respType proto.Message
-	}{
-		{
-			name: "empty chain name returns error",
-			args: []string{
-				"",
-				src.GetAddress().String(),
-			},
-			expErr:   true,
-			respType: &sdk.TxResponse{},
-		},
-		{
-			name: "empty address returns error",
-			args: []string{
-				"cosmos",
-				"",
-			},
-			expErr:   true,
-			respType: &sdk.TxResponse{},
-		},
-		{
-			name: "valid request works properly",
-			args: []string{
-				"cosmos",
-				src.GetAddress().String(),
-				fmt.Sprintf("--%s=%s", flags.FlagFrom, "dest"),
-				fmt.Sprintf("--%s=true", flags.FlagSkipConfirmation),
-				fmt.Sprintf("--%s=%s", flags.FlagBroadcastMode, flags.BroadcastBlock),
-				fmt.Sprintf("--%s=%s", flags.FlagFees, sdk.NewCoins(sdk.NewCoin(s.cfg.BondDenom, sdk.NewInt(10))).String()),
-			},
-			expErr:   false,
-			respType: &sdk.TxResponse{},
-		},
-	}
-
-	for _, tc := range testCases {
-		tc := tc
-
-		s.Run(tc.name, func() {
-			cmd := cli.GetCmdUnlinkChainAccount()
-			out, err := clitestutil.ExecTestCLICmd(cliCtx, cmd, tc.args)
-
-			if tc.expErr {
-				s.Require().Error(err)
-			} else {
-				s.Require().NoError(err)
-				s.Require().NoError(cliCtx.JSONMarshaler.UnmarshalJSON(out.Bytes(), tc.respType), out.String())
-			}
-		})
-	}
+func (s *IntegrationTestSuite) TearDownSuite() {
+	s.T().Log("tearing down integration test suite")
+	s.network.Cleanup()
 }
