+++ x/profiles/types/codec.go
@@ -29,7 +29,12 @@ func RegisterLegacyAminoCodec(cdc *codec.LegacyAmino) {
 func RegisterInterfaces(registry types.InterfaceRegistry) {
 	registry.RegisterImplementations((*authtypes.AccountI)(nil), &Profile{})
 	registry.RegisterImplementations((*authtypes.GenesisAccount)(nil), &Profile{})
-	registry.RegisterImplementations((*AddressData)(nil), &Bech32Address{}, &Base58Address{})
+	registry.RegisterInterface(
+		"desmos.profiles.v1beta1.AddressData",
+		(*AddressData)(nil),
+		&Bech32Address{},
+		&Base58Address{},
+	)
 
 	registry.RegisterImplementations((*sdk.Msg)(nil),
 		&MsgSaveProfile{},
