+++ x/profiles/types/models_packets_test.go
@@ -1,7 +1,6 @@
 package types_test
 
 import (
-	"fmt"
 	"testing"
 
 	"github.com/cosmos/cosmos-sdk/crypto/keys/secp256k1"
@@ -12,20 +11,27 @@ import (
 
 func TestLinkChainAccountPacketData_Validate(t *testing.T) {
 	tests := []struct {
-		name   string
-		packet types.LinkChainAccountPacketData
-		expErr error
+		name      string
+		packet    types.LinkChainAccountPacketData
+		shouldErr bool
 	}{
 		{
 			name: "Null source address returns error",
 			packet: types.LinkChainAccountPacketData{
-				nil,
-				types.NewProof(secp256k1.GenPrivKey().PubKey(), "032086ede8d4bce29fe364a94744ca71dbeaf370221ba20f9716a165c54b079561", "plain_text"),
-				types.NewChainConfig("cosmos"),
-				"cosmos1yt7rqhj0hjw92ed0948r2pqwtp9smukurqcs70",
-				types.NewProof(secp256k1.GenPrivKey().PubKey(), "032086ede8d4bce29fe364a94744ca71dbeaf370221ba20f9716a165c54b079561", "plain_text"),
+				SourceProof: types.NewProof(
+					secp256k1.GenPrivKey().PubKey(),
+					"032086ede8d4bce29fe364a94744ca71dbeaf370221ba20f9716a165c54b079561",
+					"plain_text",
+				),
+				SourceChainConfig:  types.NewChainConfig("cosmos"),
+				DestinationAddress: "cosmos1yt7rqhj0hjw92ed0948r2pqwtp9smukurqcs70",
+				DestinationProof: types.NewProof(
+					secp256k1.GenPrivKey().PubKey(),
+					"032086ede8d4bce29fe364a94744ca71dbeaf370221ba20f9716a165c54b079561",
+					"plain_text",
+				),
 			},
-			expErr: fmt.Errorf("source address cannot be nil"),
+			shouldErr: true,
 		},
 		{
 			name: "Invalid source proof returns error",
@@ -34,59 +40,96 @@ func TestLinkChainAccountPacketData_Validate(t *testing.T) {
 				types.NewProof(secp256k1.GenPrivKey().PubKey(), "=", "wrong"),
 				types.NewChainConfig("cosmos"),
 				"cosmos1yt7rqhj0hjw92ed0948r2pqwtp9smukurqcs70",
-				types.NewProof(secp256k1.GenPrivKey().PubKey(), "032086ede8d4bce29fe364a94744ca71dbeaf370221ba20f9716a165c54b079561", "plain_text"),
+				types.NewProof(
+					secp256k1.GenPrivKey().PubKey(),
+					"032086ede8d4bce29fe364a94744ca71dbeaf370221ba20f9716a165c54b079561",
+					"plain_text",
+				),
 			),
-			expErr: fmt.Errorf("failed to decode hex string of signature"),
+			shouldErr: true,
 		},
 		{
 			name: "Invalid chain config returns error",
 			packet: types.NewLinkChainAccountPacketData(
 				types.NewBech32Address("cosmos1yt7rqhj0hjw92ed0948r2pqwtp9smukurqcs70", "cosmos"),
-				types.NewProof(secp256k1.GenPrivKey().PubKey(), "032086ede8d4bce29fe364a94744ca71dbeaf370221ba20f9716a165c54b079561", "plain_text"),
+				types.NewProof(
+					secp256k1.GenPrivKey().PubKey(),
+					"032086ede8d4bce29fe364a94744ca71dbeaf370221ba20f9716a165c54b079561",
+					"plain_text",
+				),
 				types.NewChainConfig(""),
 				"cosmos1yt7rqhj0hjw92ed0948r2pqwtp9smukurqcs70",
-				types.NewProof(secp256k1.GenPrivKey().PubKey(), "032086ede8d4bce29fe364a94744ca71dbeaf370221ba20f9716a165c54b079561", "plain_text"),
+				types.NewProof(
+					secp256k1.GenPrivKey().PubKey(),
+					"032086ede8d4bce29fe364a94744ca71dbeaf370221ba20f9716a165c54b079561",
+					"plain_text",
+				),
 			),
-			expErr: fmt.Errorf("chain name cannot be empty or blank"),
+			shouldErr: true,
 		},
 		{
 			name: "Invalid destination address returns error",
 			packet: types.NewLinkChainAccountPacketData(
 				types.NewBech32Address("cosmos1yt7rqhj0hjw92ed0948r2pqwtp9smukurqcs70", "cosmos"),
-				types.NewProof(secp256k1.GenPrivKey().PubKey(), "032086ede8d4bce29fe364a94744ca71dbeaf370221ba20f9716a165c54b079561", "plain_text"),
+				types.NewProof(
+					secp256k1.GenPrivKey().PubKey(),
+					"032086ede8d4bce29fe364a94744ca71dbeaf370221ba20f9716a165c54b079561",
+					"plain_text",
+				),
 				types.NewChainConfig("cosmos"),
 				"",
-				types.NewProof(secp256k1.GenPrivKey().PubKey(), "032086ede8d4bce29fe364a94744ca71dbeaf370221ba20f9716a165c54b079561", "plain_text"),
+				types.NewProof(
+					secp256k1.GenPrivKey().PubKey(),
+					"032086ede8d4bce29fe364a94744ca71dbeaf370221ba20f9716a165c54b079561",
+					"plain_text",
+				),
 			),
-			expErr: fmt.Errorf("invalid destination address: %s", ""),
+			shouldErr: true,
 		},
 		{
 			name: "Invalid destination proof returns error",
 			packet: types.NewLinkChainAccountPacketData(
 				types.NewBech32Address("cosmos1yt7rqhj0hjw92ed0948r2pqwtp9smukurqcs70", "cosmos"),
-				types.NewProof(secp256k1.GenPrivKey().PubKey(), "032086ede8d4bce29fe364a94744ca71dbeaf370221ba20f9716a165c54b079561", "plain_text"),
+				types.NewProof(
+					secp256k1.GenPrivKey().PubKey(),
+					"032086ede8d4bce29fe364a94744ca71dbeaf370221ba20f9716a165c54b079561",
+					"plain_text",
+				),
 				types.NewChainConfig("cosmos"),
 				"cosmos1yt7rqhj0hjw92ed0948r2pqwtp9smukurqcs70",
 				types.NewProof(secp256k1.GenPrivKey().PubKey(), "=", "wrong"),
 			),
-			expErr: fmt.Errorf("failed to decode hex string of signature"),
+			shouldErr: true,
 		},
 		{
 			name: "Valid packet returns no error",
 			packet: types.NewLinkChainAccountPacketData(
 				types.NewBech32Address("cosmos1yt7rqhj0hjw92ed0948r2pqwtp9smukurqcs70", "cosmos"),
-				types.NewProof(secp256k1.GenPrivKey().PubKey(), "032086ede8d4bce29fe364a94744ca71dbeaf370221ba20f9716a165c54b079561", "plain_text"),
+				types.NewProof(
+					secp256k1.GenPrivKey().PubKey(),
+					"032086ede8d4bce29fe364a94744ca71dbeaf370221ba20f9716a165c54b079561",
+					"plain_text",
+				),
 				types.NewChainConfig("cosmos"),
 				"cosmos1yt7rqhj0hjw92ed0948r2pqwtp9smukurqcs70",
-				types.NewProof(secp256k1.GenPrivKey().PubKey(), "032086ede8d4bce29fe364a94744ca71dbeaf370221ba20f9716a165c54b079561", "plain_text"),
+				types.NewProof(
+					secp256k1.GenPrivKey().PubKey(),
+					"032086ede8d4bce29fe364a94744ca71dbeaf370221ba20f9716a165c54b079561",
+					"plain_text",
+				),
 			),
-			expErr: nil,
+			shouldErr: false,
 		},
 	}
 	for _, test := range tests {
 		test := test
 		t.Run(test.name, func(t *testing.T) {
-			require.Equal(t, test.expErr, test.packet.Validate())
+			err := test.packet.Validate()
+			if test.shouldErr {
+				require.Error(t, err)
+			} else {
+				require.NoError(t, err)
+			}
 		})
 	}
 }
