+++ x/profiles/types/msgs_chain_links_test.go
@@ -1,8 +1,6 @@
 package types_test
 
 import (
-	"encoding/hex"
-	"fmt"
 	"testing"
 
 	"github.com/cosmos/cosmos-sdk/crypto/keys/secp256k1"
@@ -10,123 +8,107 @@ import (
 
 	"github.com/desmos-labs/desmos/x/profiles/types"
 
-	"github.com/cosmos/cosmos-sdk/types/bech32"
 	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
 	"github.com/stretchr/testify/require"
 )
 
-func generateMsgLinkChainAccount(t *testing.T) (*types.MsgLinkChainAccount, types.AddressData) {
-	srcPrivKey := &secp256k1.PrivKey{Key: []byte{26, 15, 45, 205, 181, 29, 11, 13, 171, 161, 135, 61, 94, 174, 82, 9, 220, 10, 66, 180, 9, 49, 96, 179, 16, 189, 143, 132, 152, 111, 59, 30}}
-	srcPubKey := srcPrivKey.PubKey()
-	srcAddr, err := bech32.ConvertAndEncode("cosmos", srcPubKey.Address().Bytes())
-	require.NoError(t, err)
-
-	srcPlainText := srcAddr
-	srcSig, err := srcPrivKey.Sign([]byte(srcPlainText))
-	require.NoError(t, err)
-	srcSigHex := hex.EncodeToString(srcSig)
-
-	destPrivKey := &secp256k1.PrivKey{Key: []byte{25, 15, 45, 205, 181, 29, 11, 13, 171, 161, 135, 61, 94, 174, 82, 9, 220, 10, 66, 180, 9, 49, 96, 179, 16, 189, 143, 132, 152, 111, 59, 30}}
-	destPubKey := destPrivKey.PubKey()
-	destAddr, err := bech32.ConvertAndEncode("cosmos", destPubKey.Address().Bytes())
-
-	return types.NewMsgLinkChainAccount(
-		types.NewBech32Address(srcAddr, "cosmos"),
-		types.NewProof(srcPubKey, srcSigHex, srcPlainText),
-		types.NewChainConfig("cosmos"),
-		destAddr,
-	), types.NewBech32Address(srcAddr, "cosmos")
-}
+var chainAddress = types.NewBech32Address("cosmos1xmquc944hzu6n6qtljcexkuhhz76mucxtgm5x0", "cosmos")
+var pubKey, _ = sdk.GetPubKeyFromBech32(
+	sdk.Bech32PubKeyTypeAccPub,
+	"cosmospub1addwnpepq0j8zw4t6tg3v8gh7d2d799gjhue7ewwmpg2hwr77f9kuuyzgqtrw5r6wec",
+)
+var msgChainLinkAccount = types.NewMsgLinkChainAccount(
+	chainAddress,
+	types.NewProof(
+		pubKey,
+		"ad112abb30e5240c7b9d21b4cc5421d76cfadfcd5977cca262523b5f5bc759457d4aa6d5c1eb6223db104b47aa1f222468be8eb5bb2762b971622ac5b96351b5",
+		"text",
+	),
+	types.NewChainConfig("cosmos"),
+	"cosmos1u9hgsqfpe3snftr7p7fsyja3wtlmj2sgf2w9yl",
+)
 
 func TestMsgLinkChainAccount_Route(t *testing.T) {
-	msg, _ := generateMsgLinkChainAccount(t)
-	require.Equal(t, "profiles", msg.Route())
+	require.Equal(t, "profiles", msgChainLinkAccount.Route())
 }
 
 func TestMsgLinkChainAccount_Type(t *testing.T) {
-	msg, _ := generateMsgLinkChainAccount(t)
-	require.Equal(t, "link_chain_account", msg.Type())
+	require.Equal(t, "link_chain_account", msgChainLinkAccount.Type())
 }
 
 func TestMsgLinkChainAccount_ValidateBasic(t *testing.T) {
-	validMsg, srcAddr := generateMsgLinkChainAccount(t)
 	tests := []struct {
-		name     string
-		msg      *types.MsgLinkChainAccount
-		expError error
+		name      string
+		msg       *types.MsgLinkChainAccount
+		shouldErr bool
 	}{
 		{
-			name: "Empty source address returns error",
+			name: "Empty chain address returns error",
 			msg: &types.MsgLinkChainAccount{
-				nil,
-				validMsg.SourceProof,
-				validMsg.SourceChainConfig,
-				validMsg.DestinationAddress,
+				Proof:       msgChainLinkAccount.Proof,
+				ChainConfig: msgChainLinkAccount.ChainConfig,
+				Signer:      msgChainLinkAccount.Signer,
 			},
-			expError: fmt.Errorf("source address cannot be nil"),
+			shouldErr: true,
 		},
 		{
-			name: "Invalid source proof returns error",
+			name: "Invalid proof returns error",
 			msg: types.NewMsgLinkChainAccount(
-				srcAddr,
+				chainAddress,
 				types.NewProof(secp256k1.GenPrivKey().PubKey(), "=", "wrong"),
-				validMsg.SourceChainConfig,
-				validMsg.DestinationAddress,
+				msgChainLinkAccount.ChainConfig,
+				msgChainLinkAccount.Signer,
 			),
-			expError: fmt.Errorf("failed to decode hex string of signature"),
+			shouldErr: true,
 		},
 		{
 			name: "Invalid chain config returns error",
 			msg: types.NewMsgLinkChainAccount(
-				srcAddr,
-				validMsg.SourceProof,
+				chainAddress,
+				msgChainLinkAccount.Proof,
 				types.NewChainConfig(""),
-				validMsg.DestinationAddress,
+				msgChainLinkAccount.Signer,
 			),
-			expError: fmt.Errorf("chain name cannot be empty or blank"),
+			shouldErr: true,
 		},
 		{
-			name: "Invalid destination address returns error",
+			name: "Invalid signer returns error",
 			msg: types.NewMsgLinkChainAccount(
-				srcAddr,
-				validMsg.SourceProof,
-				validMsg.SourceChainConfig,
+				chainAddress,
+				msgChainLinkAccount.Proof,
+				msgChainLinkAccount.ChainConfig,
 				"",
 			),
-			expError: fmt.Errorf("invalid destination address: %s", ""),
+			shouldErr: true,
 		},
 		{
-			name:     "No error message",
-			msg:      validMsg,
-			expError: nil,
+			name:      "Valid message returns no error",
+			msg:       msgChainLinkAccount,
+			shouldErr: false,
 		},
 	}
 
 	for _, test := range tests {
 		test := test
 		t.Run(test.name, func(t *testing.T) {
-			returnedError := test.msg.ValidateBasic()
-			if test.expError == nil {
-				require.Nil(t, returnedError)
+			err := test.msg.ValidateBasic()
+			if test.shouldErr {
+				require.Error(t, err)
 			} else {
-				require.NotNil(t, returnedError)
-				require.Equal(t, test.expError.Error(), returnedError.Error())
+				require.NoError(t, err)
 			}
 		})
 	}
 }
 
 func TestMsgLinkChainAccount_GetSignBytes(t *testing.T) {
-	msg, _ := generateMsgLinkChainAccount(t)
-	actual := msg.GetSignBytes()
-	expected := `{"type":"desmos/MsgLinkChainAccount","value":{"destination_address":"cosmos1u9hgsqfpe3snftr7p7fsyja3wtlmj2sgf2w9yl","source_address":{"prefix":"cosmos","value":"cosmos1ma346arwsqpmjmkctwxa5uxdx66le3nty0jeax"},"source_chain_config":{"name":"cosmos"},"source_proof":{"plain_text":"cosmos1ma346arwsqpmjmkctwxa5uxdx66le3nty0jeax","pub_key":{"type":"tendermint/PubKeySecp256k1","value":"A7v3HEjiNO2jXJA+2gcBtO2VQ6Vsirs7GODz7dN39H7Q"},"signature":"ad112abb30e5240c7b9d21b4cc5421d76cfadfcd5977cca262523b5f5bc759457d4aa6d5c1eb6223db104b47aa1f222468be8eb5bb2762b971622ac5b96351b5"}}}`
-	require.Equal(t, expected, string(actual))
+	expected := `{"type":"desmos/MsgLinkChainAccount","value":{"chain_address":{"prefix":"cosmos","value":"cosmos1xmquc944hzu6n6qtljcexkuhhz76mucxtgm5x0"},"chain_config":{"name":"cosmos"},"proof":{"plain_text":"text","pub_key":{"type":"tendermint/PubKeySecp256k1","value":"A+RxOqvS0RYdF/NU3xSolfmfZc7YUKu4fvJLbnCCQBY3"},"signature":"ad112abb30e5240c7b9d21b4cc5421d76cfadfcd5977cca262523b5f5bc759457d4aa6d5c1eb6223db104b47aa1f222468be8eb5bb2762b971622ac5b96351b5"},"signer":"cosmos1u9hgsqfpe3snftr7p7fsyja3wtlmj2sgf2w9yl"}}`
+	require.Equal(t, expected, string(msgChainLinkAccount.GetSignBytes()))
 }
 
 func TestMsgLinkChainAccount_GetSigners(t *testing.T) {
-	msg, _ := generateMsgLinkChainAccount(t)
-	addr, _ := sdk.AccAddressFromBech32(msg.DestinationAddress)
-	require.Equal(t, []sdk.AccAddress{addr}, msg.GetSigners())
+	addr, _ := sdk.AccAddressFromBech32(msgChainLinkAccount.Signer)
+	require.Equal(t, []sdk.AccAddress{addr}, msgChainLinkAccount.GetSigners())
 }
 
 // ___________________________________________________________________________________________________________________
